+++ app/controllers/lists_controller.rb
@@ -7,6 +7,7 @@ class ListsController < ApplicationController
   
   def show
     @list = List.find(params[:id])
+    authorize @list
     @item = Item.new
     @items = @list.items
     respond_to do |format|
@@ -15,12 +16,9 @@ class ListsController < ApplicationController
     end
   end
 
-  def new
-    @list = List.new
-  end
-
   def create
     @list = current_user.lists.build(list_params)
+    authorize @list
     if @list.save
       flash[:notice] = "List was saved. Add your to-do items now."
       redirect_to list_path(@list)
@@ -32,8 +30,9 @@ class ListsController < ApplicationController
 
   def update
     @list = List.find(params[:id])
+    authorize @list
     if @list.update_attributes!(list_params)
-      flash[:notice] = "List was updated."
+      # flash[:notice] = "List was updated."
       respond_to do |format|
         format.html {redirect_to @list}
         format.json {render :json => @list}
@@ -49,6 +48,7 @@ class ListsController < ApplicationController
 
   def destroy
     @list = List.find(params[:id])
+    authorize @list
     if @list.destroy
       flash[:notice] = "List name was deleted."
       redirect_to lists_path
